requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.26
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
